maelstromd: add pullimageonstart option. implement pullimageonput support.
Both flags are false by default, which means images are only pulled if
they do not exist locally.

If pullimageonput is true, the image will be pulled immediately when the
component is modified.

If pullimageonstart is true, the image will be pulled every time before
starting a container for the component.