maelstromd: add softconcurrencylimit to component
If this is true, the maxconcurrency value will only be used to
autoscale the component, but will not be used as a hard limit by
the reverse proxy. That is, the limit may be exceeded.

This is useful for cases if a component is used in a request cycle
where a hard limit would potentially cause deadlocks.